% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.R
\name{timeToComplete}
\alias{timeToComplete}
\title{A function to take in gCode and return the distance and time it will take to traverse}
\usage{
timeToComplete(g0Code, feedRate = 120)
}
\description{
A function to take in gCode and return the distance and time it will take to traverse
}

